test: migrate #40d reentrancy tests to Swift Testing + AppLogger cleanup
Addresses Gemini Code Assist feedback on PR #82:

1. AppDelegate.swift — the `print` + `fflush` debug statements in the
   `voiceTriggerEnabledDidChange` observer (one outside the Task hop,
   one inside) are pre-existing artifacts that landed inside the diff
   when this PR rewrote the surrounding closure for the Notification
   sending fix. Replace with `AppLogger.debug(AppLogger.app, ...)` per
   the project's logging convention. Single line now combines both
   pieces of state ("notification received (enabled=...)") so we don't
   log twice for one event.

2. Move the four #40d reentrancy tests from `FluidAudioServiceTests`
   (XCTest) into a new `FluidAudioServiceReentrancyTests.swift` Swift
   Testing suite per the styleguide: new pure-logic / async tests use
   `@Test` / `@Suite` / `#expect`. Each test carries the appropriate
   `.fast` or `.slow` tag (`Tests/SpeechToTextTests/Utilities/TestTags.swift`)
   so CI can filter on the tag scheme #23 introduced. The XCTestCase
   keeps the pre-existing init / language / shutdown / error-description
   coverage; that file's existing shape and the new file's shape now
   match the styleguide split (XCTest for legacy + ViewInspector,
   Swift Testing for new logic).
