+++ Tests/SpeechToTextTests/Services/FluidAudioServiceTests.swift
@@ -265,206 +265,11 @@ final class FluidAudioServiceTests: XCTestCase {
     }
 
     // MARK: - Reentrancy Guard Tests (#40d)
-
-    /// Smoke test for the single-flight wrapper boilerplate introduced in
-    /// #40d: under contention the wrapper does not deadlock, does not
-    /// crash, and every concurrent call surfaces `notInitialized` rather
-    /// than mismatched / mixed errors. This exercises the *fast* path
-    /// (each caller acquires + releases the slot synchronously between
-    /// actor turns); the *wait* path is covered by
-    /// `test_transcribe_serialisesConcurrentCallsViaWaitPath` below.
-    func test_transcribe_concurrentCallsAllReturnNotInitialized() async {
-        // Given
-        let service = FluidAudioService()
-        let samples: [Int16] = Array(repeating: 100, count: 1600)
-        let callCount = 8
-
-        // When — fire N concurrent transcribes against an uninitialised
-        // service. Each call's `runTranscribe` will throw `notInitialized`
-        // because `asrManager` is nil.
-        let errors: [Error?] = await withTaskGroup(of: Error?.self) { group in
-            for _ in 0..<callCount {
-                group.addTask {
-                    do {
-                        _ = try await service.transcribe(
-                            samples: samples,
-                            sampleRate: 16000.0
-                        )
-                        return nil
-                    } catch {
-                        return error
-                    }
-                }
-            }
-            var collected: [Error?] = []
-            for await error in group { collected.append(error) }
-            return collected
-        }
-
-        // Then — every caller surfaced an error (none returned a value),
-        // and every error is `notInitialized` (no other failure mode).
-        // If the guard deadlocked, this would hang and fail via timeout.
-        XCTAssertEqual(errors.count, callCount)
-        for case let .some(error) in errors {
-            guard let fluidError = error as? FluidAudioError else {
-                XCTFail("Expected FluidAudioError, got \(type(of: error))")
-                continue
-            }
-            XCTAssertEqual(fluidError, .notInitialized)
-        }
-    }
-
-    /// Forces the *wait* path of the single-flight guard: the holder is
-    /// kept in flight long enough that subsequent callers genuinely
-    /// queue on `transcribeWaiters`. Verifies FIFO hand-off works (every
-    /// caller gets resumed) without a missed `resume()` (which would
-    /// hang the test).
-    ///
-    /// Uses the `transcribeSimulatedDelay` test seam to slow the active
-    /// transcribe so reentrant callers observe `transcribeInFlight ==
-    /// true` and suspend on a continuation. Combined with `simulatedError
-    /// = .transcription`, every call surfaces `transcriptionFailed`
-    /// cleanly — proving the queue drained and no waiter was stranded.
-    func test_transcribe_serialisesConcurrentCallsViaWaitPath() async {
-        // Given — service that holds the slot for 50ms per call,
-        // forcing reentrant callers to queue.
-        let service = FluidAudioService(
-            simulatedError: .transcription,
-            transcribeSimulatedDelay: .milliseconds(50)
-        )
-        let samples: [Int16] = Array(repeating: 100, count: 1600)
-        let callCount = 5
-        let started = Date()
-
-        // When — fire N concurrent calls; the first acquires the slot,
-        // the rest queue.
-        let errors: [Error?] = await withTaskGroup(of: Error?.self) { group in
-            for _ in 0..<callCount {
-                group.addTask {
-                    do {
-                        _ = try await service.transcribe(
-                            samples: samples,
-                            sampleRate: 16000.0
-                        )
-                        return nil
-                    } catch {
-                        return error
-                    }
-                }
-            }
-            var collected: [Error?] = []
-            for await error in group { collected.append(error) }
-            return collected
-        }
-        let elapsed = Date().timeIntervalSince(started)
-
-        // Then — every caller threw, every error is the simulated
-        // transcriptionFailed (proves runTranscribe ran on each, slot
-        // released cleanly between).
-        XCTAssertEqual(errors.count, callCount)
-        for case let .some(error) in errors {
-            guard let fluidError = error as? FluidAudioError else {
-                XCTFail("Expected FluidAudioError, got \(type(of: error))")
-                continue
-            }
-            if case .transcriptionFailed = fluidError {
-                // pass
-            } else {
-                XCTFail("Expected .transcriptionFailed, got \(fluidError)")
-            }
-        }
-
-        // And — total elapsed time should be >= callCount * delay
-        // (modulo scheduling jitter). If the calls had run *concurrently*
-        // (guard broken), elapsed would be ~one delay. We use a forgiving
-        // lower bound (75%) to absorb test-runner jitter while still
-        // catching a fully-broken guard.
-        let lowerBound = Double(callCount) * 0.050 * 0.75
-        XCTAssertGreaterThanOrEqual(
-            elapsed, lowerBound,
-            "expected serialised execution (~\(callCount) × 50ms), got \(elapsed)s"
-        )
-    }
-
-    /// After concurrent calls all complete, the service must be ready to
-    /// accept another call — i.e. the in-flight slot was released. This
-    /// catches a class of bug where the slot is handed to a waiter on the
-    /// failure path but the flag never clears (next caller would hang).
-    func test_transcribe_isUsableAfterContention() async {
-        // Given
-        let service = FluidAudioService()
-        let samples: [Int16] = Array(repeating: 100, count: 1600)
-
-        // When — drive contention then issue one more call.
-        await withTaskGroup(of: Void.self) { group in
-            for _ in 0..<4 {
-                group.addTask {
-                    _ = try? await service.transcribe(samples: samples, sampleRate: 16000.0)
-                }
-            }
-        }
-
-        // Then — a follow-up call still throws cleanly (no hang, no crash).
-        do {
-            _ = try await service.transcribe(samples: samples, sampleRate: 16000.0)
-            XCTFail("Should still throw notInitialized")
-        } catch let error as FluidAudioError {
-            XCTAssertEqual(error, .notInitialized)
-        } catch {
-            XCTFail("Wrong error type: \(error)")
-        }
-    }
-
-    /// Verifies that `shutdown()` drains queued waiters so their
-    /// `CheckedContinuation`s don't leak and their callers don't hang
-    /// forever. Uses the wait-path delay seam to ensure waiters are
-    /// genuinely queued before shutdown.
-    func test_shutdown_drainsQueuedWaiters() async {
-        // Given — service set up so the active transcribe blocks long
-        // enough for a second caller to queue, then we shut down.
-        let service = FluidAudioService(
-            simulatedError: .transcription,
-            transcribeSimulatedDelay: .milliseconds(200)
-        )
-        let samples: [Int16] = Array(repeating: 100, count: 1600)
-
-        // When — kick off the holder, give it a moment to acquire the
-        // slot, kick off a second caller (which will queue), then shut
-        // down before the holder completes.
-        async let holder: Error? = {
-            do {
-                _ = try await service.transcribe(samples: samples, sampleRate: 16000.0)
-                return nil
-            } catch {
-                return error
-            }
-        }()
-        // Yield so the holder reaches the slot acquisition.
-        try? await Task.sleep(for: .milliseconds(20))
-
-        async let queued: Error? = {
-            do {
-                _ = try await service.transcribe(samples: samples, sampleRate: 16000.0)
-                return nil
-            } catch {
-                return error
-            }
-        }()
-        // Yield so the second caller reaches the wait path.
-        try? await Task.sleep(for: .milliseconds(20))
-
-        await service.shutdown()
-
-        // Then — both calls return (no hang). The holder finishes its
-        // simulated delay then throws `transcriptionFailed`. The queued
-        // caller is woken by shutdown's drain, hits `notInitialized`
-        // because `asrManager` is now nil.
-        let holderError = await holder
-        let queuedError = await queued
-
-        XCTAssertNotNil(holderError, "holder should have surfaced an error")
-        XCTAssertNotNil(queuedError, "queued caller should have surfaced an error")
-        // Holder ran through its body; queued was drained post-shutdown.
-        // Either error type is acceptable as long as both paths terminated.
-    }
+    //
+    // The reentrancy / single-flight tests for #40d live in the
+    // sibling Swift Testing suite
+    // `FluidAudioServiceReentrancyTests.swift` (per the styleguide:
+    // new pure-logic / async tests use `@Test` / `#expect`, tagged
+    // `.fast` / `.slow`). This XCTestCase remains for the pre-existing
+    // init / language / shutdown / error-description coverage.
 }
